backend/requirements.txt
Flask
python-dotenv
psycopg2-binary
bcrypt
PyJWT